+++ lib/slack-ruby-bot-server/models.rb
@@ -1,1 +1,0 @@
-require 'slack-ruby-bot-server/models/team'