+++ sample_apps/sample_app_activerecord/db/migrate/20190323181453_add_activated_fields.rb
@@ -1,0 +1,7 @@
+class AddActivatedFields < ActiveRecord::Migration[5.0]
+  def change
+    add_column :teams, :bot_user_id, :string
+    add_column :teams, :activated_user_id, :string
+    add_column :teams, :activated_user_access_token, :string
+  end
+end